// src/app/core/enums/order-status.ts
export enum OrderStatus {
  CREATED = 'CREATED',
  DELIVERED = 'DELIVERED',
  FAILED = 'FAILED',

  REGISTERED = 'REGISTERED',
  PREPARING = 'PREPARING',
  TRANSIT = 'TRANSIT',
  DELIVERED_CLT = 'DELIVERED_CLT',
  RETURNED = 'RETURNED',
  BLOCKED = 'BLOCKED',
  CANCELLED = 'CANCELLED',
  PARTIAL_DELIVERY = 'PARTIAL_DELIVERY',
  ORDER_NOT_PROCESSED = 'ORDER_NOT_PROCESSED',
}